Moved more types to their own files.
